[#135] refactor : 폴더이름 변경(accounts를 account)로
